LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-stretch-properties-001-expected.txt


FAIL Stretchy vertical line assert_approx_equals: Infix <mo> element expected 25 +/- 1 but got 150
|
|
|
|
